+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Updated some dependencies. Now Nekomata can be built with GHC 9.8.1.
 * Fixed a bug in printing truncated results.
 * Now characters are internally represented as `Word8`s instead of `Char`s.
+* Added a REPL command `\Arity` to print the arity of a function.
+* New built-in functions: `\minMax`, `\pad`.
 
 ### Breaking changes